% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\name{affine.map}
\alias{affine.map}
\title{affine.map}
\usage{
affine.map(x, ylim = c(0, 1), xlim = c(min(x), max(x)), cap = FALSE,
  cap.min = cap, cap.max = cap)

affine.map(x, ylim = c(0, 1), xlim = c(min(x), max(x)), cap = FALSE,
  cap.min = cap, cap.max = cap)
}
\description{
affinely maps 1D points in vector x from interval xlim to interval ylim,
ie takes points that lie in
interval xlim and mapping onto interval ylim using linear / affine map defined by:
(x0,y0) = c(xlim(1), ylim(1)),
(x1,y1) = c(xlim(2), ylim(2))
(using two point formula for line)
useful for plotting.

if cap.max or cap.min == T then values outside of the range will be capped at min or max

affinely maps 1D points in vector x from interval xlim to interval ylim,
ie takes points that lie in
interval xlim and mapping onto interval ylim using linear / affine map defined by:
(x0,y0) = c(xlim(1), ylim(1)),
(x1,y1) = c(xlim(2), ylim(2))
(using two point formula for line)
useful for plotting.
}
\details{
if cap.max or cap.min == TRUE then values outside of the range will be capped at min or max
}
\author{
Marcin Imielinski
}
\keyword{internal}

