% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\name{alpha}
\alias{alpha}
\title{Give transparency value to colors}
\usage{
alpha(col, alpha)
}
\arguments{
\item{col}{RGB color}
}
\description{
Takes provided colors and gives them the specified alpha (ie transparency) value
}
\keyword{internal}

