% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\name{barbs}
\alias{barbs}
\title{barbs}
\usage{
barbs(x0, y0, x1, y1, angle = 0, ...)
}
\description{
(improvement over pencils)
used to represent directional ranges.  Shape consists of two horizontally reflected parallelograms
with a specified angle.  The middle segment of the parallelogram represents the provided range.

Parameterized by x0, x1 (range values), y0, y1 (thickness), angle, where angle is between -90 and 90

eg angle of 0 = rectangle
   angle of 45 = rightward pointing barb
   angle of -45 = leftward pointing barb

Note: the shape will NOT be necessarily contained inside the box outlined by x0, x1, y0, y1
}
\author{
Marcin Imielinski
}
\keyword{internal}
