% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\name{bernsteinp}
\alias{bernsteinp}
\title{bernsteinp}
\usage{
bernsteinp(n, m)
}
\description{
bernsteinp

computes matrix of bernstein polynomials of degree n at m points t in the
interval 0, 1

out[i,j] = choose(n, j)*t[i]^j*(1-t[i])^(n-j)
where t = seq(0, 1, length.out = m) and j in 0 .. n

this matrix can be multiplied by a n x 2 matrix of control points to yield
an m x 2 matrix of m equally spaced points along the curve
}
\author{
Marcin Imielinski
}
\keyword{internal}
