% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\name{brewer.master}
\alias{brewer.master}
\title{brewer.master}
\usage{
brewer.master(n, palette = "Accent")
}
\arguments{
\item{palette}{any brewer.pal palette to begin with (default 'Accent')}

\item{scalar}{positive integer of number of colors to return}
}
\description{
Make brewer colors using entire palette

Makes a lot of brewer colors using entire brewer palette
}
\author{
Marcin Imielinski
}
\keyword{internal}

