% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{c}
\alias{c}
\alias{c,gTrack-method}
\title{c}
\usage{
\S4method{c}{gTrack}(x, ..., recursive = FALSE)
}
\description{
Concatenate gTrack objects gt1, gt2, gt3
usage:
c(gt1, gt2, gt3) # returns a gTrack object with the component tracks "stacked"
}
\author{
Marcin Imielinski
}

