% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,gTrack-method}
\title{$}
\usage{
\S4method{$}{gTrack}(x, name)
}
\arguments{
\item{x}{\code{gTrack} object}

\item{name}{Name of the \code{formatting} field to view}
}
\description{
Accessing columns of gTrack formatting data.frame
}
\author{
Marcin Imielinski
}
