% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{$<-}
\alias{$<-}
\title{$<-}
\usage{
\S4method{$}{gTrack}(x, name) <- value
}
\description{
Setting formats of gTrack object - ie modifying the formatting(gt) data frame after
an object has already been instantiated

Usage:
gt$y.field = 'score'
gt$gr.colorfield[1] = 'readtype'
}
\author{
Marcin Imielinski
}

