% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{clear}
\alias{clear}
\alias{clear,gTrack-method}
\title{clear}
\usage{
clear(.Object)

\S4method{clear}{gTrack}(.Object)
}
\arguments{
\item{.Object}{\code{gTrack} object to clear}
}
\description{
Clear data from gTrack object (makes into an empty container with the same seqinfo and display features)
usage:
clear(gTrack)
}
\author{
Marcin Imielinski
}
