% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{clear}
\alias{clear}
\title{clear}
\usage{
\S4method{clear}{gTrack}(.Object)
}
\description{
Clear data from gTrack object (makes into an empty container with the same seqinfo and display features)
usage:
clear(gTrack)
}
\author{
Marcin Imielinski
}

