% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{colormap<-}
\alias{colormap<-}
\alias{colormap<-,gTrack-method}
\title{colormap<-}
\usage{
colormap(.Object) <- value

\S4method{colormap}{gTrack}(.Object) <- value
}
\arguments{
\item{.Object}{\code{gTrack} to set the colormap for}

\item{value}{New \code{colormap}}
}
\description{
Set colormap of gTrack object, this is a named list of named character vectors that specifies
the field of the underlying GRanges object that will be used to map a set of values
to a set of colors.

usage:
colormap(gt)[1] = list(tumortype = c(lung = 'red', pancreatic = 'blue', colon = 'purple'))
}
