% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{colormap}
\alias{colormap}
\title{colormap}
\usage{
\S4method{colormap}{gTrack}(.Object)

\S4method{colormap}{gTrack}(.Object) <- value
}
\description{
Access colormap of gTrack object, this is a named list of named character vectors that specifies
the field of the underlying GRanges object that will be used to map a set of values
to a set of colors.
usage:
colormap(gt)

Set colormap of gTrack object, this is a named list of named character vectors that specifies
the field of the underlying GRanges object that will be used to map a set of values
to a set of colors.

usage:
colormap(gt)[1] = list(tumortype = c(lung = 'red', pancreatic = 'blue', colon = 'purple'))
}
\author{
Marcin Imielinski

Marcin Imielinski
}

