% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{dat}
\alias{dat}
\alias{dat,gTrack-method}
\title{dat}
\usage{
dat(.Object)

\S4method{dat}{gTrack}(.Object)
}
\arguments{
\item{.Object}{\code{gTrack} to retrive data from}
}
\description{
Extract list of data contained inside a gTrack object gt.  Each list item will contain
either a GRanges, RleList, or path to a file.

usage
dat(gt)
}
\author{
Marcin Imielinski
}
