% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.utils.R
\name{dedup}
\alias{dedup}
\title{Deduplicate character vectors
Relabels duplicates in a character vector with .1, .2, .3
(where "." can be replaced by any user specified suffix)}
\usage{
dedup(x, suffix = ".")
}
\arguments{
\item{x}{Character vector to deduplicate.}

\item{suffix}{User defined suffix}
}
\value{
dedupclicated character vector
}
\description{
Deduplicate character vectors
Relabels duplicates in a character vector with .1, .2, .3
(where "." can be replaced by any user specified suffix)
}
\keyword{internal}

