% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\name{draw.ranges}
\alias{draw.ranges}
\title{draw.ranges}
\usage{
draw.ranges(x, y = NULL, lwd = 0.5, col = "black", border = col,
  label = NULL, draw.backbone = F, group = NULL, col.backbone = col[1],
  cex.label = 1, srt.label = 45, adj.label = c(0.5, 0), angle,
  circles = FALSE, bars = FALSE, points = NA, lines = FALSE,
  y0.bar = NULL, strand.dir = T, xlim = NULL, ylim = NULL,
  clip = NULL, lwd.border = 1, lwd.backbone = 1, verbose = FALSE, ...)
}
\description{
Draws (genomic, Iranges, Rle, or df with start and end field) intervals
as rectangles of thickness "lwd", col "col".

If "col" or "y" are set to null then will examine "x" to see if the respective fields are contained within (e.g. as RangedData or data.frame)

Adds optional backbone between beginning of first range
and end of last on existing plot
}
\author{
Marcin Imielinski
}
\keyword{internal}

