% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\name{draw.triangle}
\alias{draw.triangle}
\title{draw.triangle}
\usage{
draw.triangle(
  grl,
  mdata,
  y,
  ylim.parent = NULL,
  windows = NULL,
  win.gap = NULL,
  m.bg.col = NA,
  sigma = NA,
  col.min = "white",
  col.max = "red",
  gr.colormap = NA,
  cmap.min = NA,
  cmap.max = NA,
  m.sep.lwd = NA,
  legend = TRUE,
  max.ranges = Inf,
  leg.params = list(),
  min.gapwidth = 1,
  islog = FALSE,
  if.flip = FALSE
)
}
\description{
internal draw.triangle function
}
\author{
Jeremiah Wala
}
\keyword{12:27:43}
\keyword{2019}
\keyword{28,}
\keyword{AM}
\keyword{Jan}
\keyword{Monday,}
\keyword{internal}
\keyword{mimielinski}
