% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{edgs}
\alias{edgs}
\alias{edgs,gTrack-method}
\title{edgs}
\usage{
edgs(.Object)

\S4method{edgs}{gTrack}(.Object)
}
\arguments{
\item{.Object}{\code{gTrack} object to extract the edge list from}
}
\description{
Get edges list associated with gTrack object, is is a list of data.frames
each which contains a data.frame specifying formatted edges, as pairs of indices into the
underlying GRanges object $from, $to and additional formats $col $lwd $lty

usage:
edgs(gt)
}
