% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{edgs}
\alias{edgs}
\title{edgs}
\usage{
\S4method{edgs}{gTrack}(.Object)
}
\description{
Get edges list associated with gTrack object, is is a list of data.frames
each which contains a data.frame specifying formatted edges, as pairs of indices into the
underlying GRanges object $from, $to and additional formats $col $lwd $lty

usage:
edgs(gt)
}
\author{
Marcin Imielinski
}

