% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{formatting}
\alias{formatting}
\alias{formatting,gTrack-method}
\title{formatting}
\usage{
formatting(.Object, ...)

\S4method{formatting}{gTrack}(.Object)
}
\description{
Get data frame specifying formatting of gTrack object gt
usage:
formatting(gt)
gt # will just display the formatting data.frame

If you want to access particular fields of the formatting data.frame, just use the "$" accessor like you would for a data.frame
}
\author{
Marcin Imielinski
}

