% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{formatting<-}
\alias{formatting<-}
\alias{formatting<-,gTrack-method}
\title{formatting<-}
\usage{
formatting(.Object) <- value

\S4method{formatting}{gTrack}(.Object) <- value
}
\arguments{
\item{.Object}{\code{gTrack} to set the formatting field for}

\item{value}{\code{data.frame} with the new formatting information}
}
\description{
Set formatting of gTrack object  gt

usage:
#gt is a gTrack object
formatting(td)$height <- 2

#can also just use $ directly, like for a data frame, which is more convenient
td$height <- 2
}
