% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{formatting<-}
\alias{formatting<-}
\title{formatting<-}
\usage{
\S4method{formatting}{gTrack}(.Object) <- value
}
\description{
Set formatting of gTrack object  gt

usage:
#gt is a gTrack object
formatting(td)$height <- 2

#can also just use $ directly, like for a data frame, which is more convenient
td$height <- 2
}
\author{
Marcin Imielinski
}

