% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_gaxis.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,gaxis-method}
\title{$}
\usage{
\S4method{$}{gaxis}(x, name)
}
\arguments{
\item{x}{\code{gaxis} object}

\item{name}{Name of the \code{formatting} field to view}
}
\description{
Accessing columns of gaxis formatting data.frame
}
\author{
Marcin Imielinski
}

