% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_gaxis.R
\docType{class}
\name{gaxis-class}
\alias{gaxis}
\alias{gaxis-class}
\title{S4 class for x-axis params of \code{gTrack}

Class \code{gaxis} is a simple container storing x-axis plotting information for gTrack}
\usage{
gaxis(tick.adjust.x = 1.0001, tick.adjust.y = 0.50001,
  label.adjust.x = 0.5, label.adjust.y = 0, prefix = "", unit = 1,
  suffix = "", round = 3, cex.label = 1, newline = TRUE,
  chronly = FALSE, width = TRUE, interval = "auto", label.angle = 0,
  ticklen = 1, cex.tick = 1, min = NA, max = NA, min.bar = NA,
  pretty = TRUE, plot = TRUE, quantile = 0.01, cap = TRUE,
  tick.angle = 90, log = FALSE, breaks = NA, labels = NA)
}
\arguments{
\item{unit}{vector or scalar numeric specifying the unit that will be used in describing x axis coordinates (TODO: move to display)  (formatting)}

\item{suffix}{vector or scalar numeric specifying the suffix that will be used in describin x axis coordinates (TODO: move to display) (formatting)}

\item{round}{vector or scalar non-neg integer specifying number of decimals to round axis coordinate labels (formatting)}

\item{newline}{vector or scalar logical specifying whether to draw a newline in the axis coordinate labels (formatting)}

\item{width}{Logical scalar specifying whether to add window width to axis window labels [TRUE]}

\item{label.angle}{vector or scalar numeric between 0 and 360 specifying angle with which to draw axis coordinate labels (formatting)}

\item{ticklen}{Scalar numeric specifying lengths for axis ticks  (formatting)}

\item{cex.tick}{Scalar numeric specifying expansion factor for axis tick labels  (formatting)}

\item{quantile}{if \code{min} or \code{max} is not specified then will draw between quantile and 1-quantile of data}

\item{cap}{whether to cap values at min, min (only relevant if y.field specified)
Construct a new \code{gaxis}}

\item{nticks}{vector or scalar positive integer specifying how many axis ticks to optimally draw (formatting)}
}

