% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_gaxis.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,gaxis-method}
\title{show}
\usage{
\S4method{show}{gaxis}(object)
}
\arguments{
\item{object}{\code{gaxis} to display}
}
\description{
Display a \code{gaxis} object
}
\author{
Marcin Imielinski
}

