% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glabel.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,glabel-method}
\title{$}
\usage{
\S4method{$}{glabel}(x, name)
}
\arguments{
\item{x}{\code{glabel} object}

\item{name}{Name of the \code{formatting} field to view}
}
\description{
Accessing columns of glabel formatting data.frame
}

