% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glabel.R
\docType{methods}
\name{$<-}
\alias{$<-}
\alias{$<-,glabel-method}
\title{$<-}
\usage{
\S4method{$}{glabel}(x, name) <- value
}
\arguments{
\item{x}{\code{glabel} object to alter \code{formatting} field of}

\item{name}{\code{formatting} field to alter}

\item{value}{New value}
}
\description{
Setting formats of glabel object - ie modifying the formatting(gt) data frame after
an object has already been instantiated
}

