% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glabel.R
\docType{class}
\name{glabel-class}
\alias{glabel}
\alias{glabel-class}
\title{S4 class for legend params of \code{gTrack}

Class \code{glabel} is a simple container storing legend plotting information for gTrack}
\usage{
glabel(gr.cex = 1, grl.cex = 1, grl.labelfield = NA, gr.labelfield = NA,
  gr.angle = 0, grl.angle = 0, hadj = 1, vadj = 0.5, ...)
}
\description{
Construct a new \code{glabel}
}

