% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glabel.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,glabel-method}
\title{show}
\usage{
\S4method{show}{glabel}(object)
}
\arguments{
\item{object}{\code{glabel} to display}
}
\description{
Display a \code{glabel} object
}

