% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glegend.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,glegend-method}
\title{$}
\usage{
\S4method{$}{glegend}(x, name)
}
\arguments{
\item{x}{\code{glegend} object}

\item{name}{Name of the \code{formatting} field to view}
}
\description{
Accessing columns of glegend formatting data.frame
}
\author{
Marcin Imielinski
}

