% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glegend.R
\docType{methods}
\name{$<-}
\alias{$<-}
\alias{$<-,glegend-method}
\title{$<-}
\usage{
\S4method{$}{glegend}(x, name) <- value
}
\arguments{
\item{x}{\code{glegend} object to alter \code{formatting} field of}

\item{name}{\code{formatting} field to alter}

\item{value}{New value}
}
\description{
Setting formats of glegend object - ie modifying the formatting(gt) data frame after
an object has already been instantiated
}

