% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glegend.R
\docType{class}
\name{glegend-class}
\alias{glegend}
\alias{glegend-class}
\title{S4 class for legend params of \code{gTrack}

Class \code{glegend} is a simple container storing legend plotting information for gTrack}
\usage{
glegend(xpos = 0, ypos = 1, plot = TRUE, title = "", maxitems = Inf,
  x = xpos, y = ypos, legend = NA, ...)
}
\arguments{
\item{maxitems}{Scalar positive integer specifying
what is the maximum number of items to include in legend [Inf]}
}
\description{
Construct a new \code{glegend}
}

