% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glegend.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,glegend-method}
\title{show}
\usage{
\S4method{show}{glegend}(object)
}
\arguments{
\item{object}{\code{glegend} to display}
}
\description{
Display a \code{glegend} object
}
\author{
Marcin Imielinski
}

