% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpath.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,gpath-method}
\title{$}
\usage{
\S4method{$}{gpath}(x, name)
}
\arguments{
\item{x}{\code{gpath} object}

\item{name}{Name of the \code{formatting} field to view}
}
\description{
Accessing columns of gpath formatting data.frame
}
\author{
Marcin Imielinski
}

