% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpath.R
\docType{class}
\name{gpath-class}
\alias{gpath}
\alias{gpath-class}
\title{S4 class for legend params of \code{gTrack}

Class \code{gpath} is a simple container storing legend plotting information for gTrack}
\usage{
gpath(cex.v = 1, cex.h = 1, cex.arrow = 1, col = "black",
  col.arrow = "black", stack.y.gap = 1, stack.x.gap = 0, ...)
}
\arguments{
\item{cex.v}{scalar numeric > 0 specifying vertical bulge of sline in paths}

\item{cex.h}{scalar numeric > 0 specifying horizontal bulge of spline in paths}

\item{col}{scalar character specifying color of path (only applicable for tracks with draw.paths = TRUE)}

\item{col.arrow}{scalar character specifying color of arrow of path}

\item{stack.y.gap}{scalar numeric > 0 specifying y stack gap of paths}

\item{stack.x.gap}{scalar numeric > 0 specifying x stack gap for paths}

\item{arrow}{scalar numeric > 0 specifying expansion factor of arrow of path}
}
\description{
Construct a new \code{gpath}
}

