% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpath.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,gpath-method}
\title{show}
\usage{
\S4method{show}{gpath}(object)
}
\arguments{
\item{object}{\code{gpath} to display}
}
\description{
Display a \code{gpath} object
}
\author{
Marcin Imielinski
}

