% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\name{gr.flatmap}
\alias{gr.flatmap}
\title{gr.flatmap}
\usage{
gr.flatmap(gr, windows, gap = 0, strand.agnostic = TRUE, squeeze = FALSE,
  xlim = c(0, 1))
}
\arguments{
\item{gr}{\code{GRanges} pile to flatten}

\item{windows}{\code{GRanges} pile of windows defining the coordinate system}

\item{gap}{TODO}

\item{strand.agnostic}{TODO}

\item{squeeze}{TODO}

\item{xlim}{TODO}
}
\value{
TODO
}
\description{
Takes \code{GRanges} pile and maps onto a flattened coordinate system defined by windows (GRanges object)
a provided "gap" (in sequence units).  If squeeze == TRUE then will additionally squeeze ranges into xlim.
}
\details{
output is list with two fields corresponding to data frames:
$grl.segs = data frame of input gr's "lifted" onto new flattened coordinate space (NOTE: nrow of this not necessarily equal to length(gr))
$window.segs = the coordinates of input windows in the new flattened (and squeezed) space
}
\keyword{internal}

