% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\name{gr.sub}
\alias{gr.sub}
\title{gr.sub}
\usage{
gr.sub(gr, a = c("(^chr)(\\\\.1$)", "MT"), b = c("", "M"))
}
\description{
will apply gsub to seqlevels of gr, by default removing 'chr', and "0.1" suffixes, and replacing "MT" with "M"
}
\keyword{internal}

