% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_gsep.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,gsep-method}
\title{$}
\usage{
\S4method{$}{gsep}(x, name)
}
\arguments{
\item{x}{\code{gsep} object}

\item{name}{Name of the \code{formatting} field to view}
}
\description{
Accessing columns of gsep formatting data.frame
}

