% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_gsep.R
\docType{methods}
\name{$<-}
\alias{$<-}
\alias{$<-,gsep-method}
\title{$<-}
\usage{
\S4method{$}{gsep}(x, name) <- value
}
\arguments{
\item{x}{\code{gsep} object to alter \code{formatting} field of}

\item{name}{\code{formatting} field to alter}

\item{value}{New value}
}
\description{
Setting formats of gsep object - ie modifying the formatting(gt) data frame after
an object has already been instantiated
}

