% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_gsep.R
\docType{class}
\name{gsep-class}
\alias{gsep}
\alias{gsep-class}
\title{S4 class for legend params of \code{gTrack}

Class \code{gsep} is a simple container storing legend plotting information for gTrack}
\usage{
gsep(plot = TRUE, lty = 1, lwd = 1, bg.col = "gray95", ...)
}
\arguments{
\item{plot}{Logical allowing separators to be turned off [FALSE]
what is the maximum number of items to include in legend [Inf]}

\item{lty}{Scalar integer specifying line style for window separators [2] (dashed line)}

\item{lwd}{Scalar numeric specifying line thickness for window separators [1]}

\item{bg.col}{Color (supplied by name or hex code) for the background of the windows [gray95]}
}
\description{
Construct a new \code{gsep}
}

