% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_gsep.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,gsep-method}
\title{show}
\usage{
\S4method{show}{gsep}(object)
}
\arguments{
\item{object}{\code{gsep} to display}
}
\description{
Display a \code{gsep} object
}

