% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{edgs<-,gTrack-method}
\alias{clear,gTrack-class}
\alias{clear,gTrack-method}
\alias{colormap,gTrack-class}
\alias{colormap,gTrack-method}
\alias{colormap<-,gTrack-method}
\alias{edgs<-,gTrack-class}
\alias{edgs<-,gTrack-method}
\title{Set edges data.frame associated with associated with gTrack object, the data.frame
that is being used to replace must have fields $from, $to, and can have optional fields $lwd
$lty, $col specifyign color and line type.}
\usage{
\S4method{edgs}{gTrack}(.Object) <- value

\S4method{clear}{gTrack}(.Object)

\S4method{colormap}{gTrack}(.Object)

\S4method{colormap}{gTrack}(.Object) <- value
}
\description{
usage:
edgs(gt)[[1]] <- new.edges.

Clear data from gTrack object (makes into an empty container with the same seqinfo and display features)
usage:
clear(gTrack)

Access colormap of gTrack object, this is a named list of named character vectors that specifies
the field of the underlying GRanges object that will be used to map a set of values
to a set of colors.
usage:
colormap(gt)

usage:
colormap(gt)[1] = list(tumortype = c(lung = 'red', pancreatic = 'blue', colon = 'purple'))
}

