% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_xaxis.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,gxaxis-method}
\title{$}
\usage{
\S4method{$}{gxaxis}(x, name)
}
\arguments{
\item{x}{\code{gxaxis} object}

\item{name}{Name of the \code{formatting} field to view}
}
\description{
Accessing columns of gxaxis formatting data.frame
}
\author{
Marcin Imielinski
}

