% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_xaxis.R
\docType{class}
\name{gxaxis-class}
\alias{gxaxis}
\alias{gxaxis-class}
\title{S4 class for x-axis params of \code{gTrack}

Class \code{gxaxis} is a simple container storing x-axis plotting information for gTrack}
\usage{
gxaxis(xaxis.prefix = "", xaxis.unit = 1, xaxis.suffix = "",
  xaxis.round = 3, xaxis.cex.label = 1, xaxis.newline = TRUE,
  xaxis.chronly = FALSE, xaxis.width = TRUE, xaxis.interval = "auto",
  xaxis.label.angle = 0, xaxis.ticklen = 1, xaxis.cex.tick = 1)
}
\arguments{
\item{xaxis.unit}{vector or scalar numeric specifying the unit that will be used in describing x axis coordinates (TODO: move to display)  (formatting)}

\item{xaxis.suffix}{vector or scalar numeric specifying the suffix that will be used in describin x axis coordinates (TODO: move to display) (formatting)}

\item{xaxis.round}{vector or scalar non-neg integer specifying number of decimals to round xaxis coordinate labels (formatting)}

\item{xaxis.newline}{vector or scalar logical specifying whether to draw a newline in the xaxis coordinate labels (formatting)}

\item{xaxis.width}{Logical scalar specifying whether to add window width to xaxis window labels [TRUE]}

\item{xaxis.label.angle}{vector or scalar numeric between 0 and 360 specifying angle with which to draw xaxis coordinate labels (formatting)}

\item{xaxis.ticklen}{Scalar numeric specifying lengths for axis ticks  (formatting)
Construct a new \code{gxaxis}}

\item{xaxis.cex.tick}{Scalar numeric specifying expansion factor for axis tick labels  (formatting)}

\item{xaxis.nticks}{vector or scalar positive integer specifying how many xaxis ticks to optimally draw (formatting)}
}

