% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_xaxis.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,gxaxis-method}
\title{show}
\usage{
\S4method{show}{gxaxis}(object)
}
\arguments{
\item{object}{\code{gxaxis} to display}
}
\description{
Display a \code{gxaxis} object
}
\author{
Marcin Imielinski
}

