% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\name{karyogram}
\alias{karyogram}
\alias{td.karyogram,gTrack-class}
\title{Returns gTrack displaying colormap for ucsc hg18 or hg19 karyotype}
\usage{
karyogram(hg19 = T, bands = T, arms = T, tel.width = 2e+06, ...)
}
\arguments{
\item{hg19}{logical scalar flag, if T returns gTrack for hg19}

\item{bands}{logical scalar, if T returns gTrack with colored giemsa bands}

\item{arms}{logical scalar, if T and bands F, returns chromosome arms with different colors and centromeres and telomeres marked, if arms is F and bands F returns chromosomes, each with a different color}

\item{tel.width}{numeric scalar, specifies telomere width in bases (only relevant if arms = T, bands = F)}
}
\description{
Returns gTrack displaying colormap for ucsc hg18 or hg19 karyotype
}

