% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{length,gTrack-method}
\alias{length,gTrack-class}
\alias{length,gTrack-method}
\title{Getting length of gTrack object gt
Usage:
length(gt)}
\usage{
\S4method{length}{gTrack}(x)
}
\description{
Getting length of gTrack object gt
Usage:
length(gt)
}

