% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{length}
\alias{length}
\alias{length,gTrack-method}
\title{length}
\usage{
\S4method{length}{gTrack}(x)
}
\arguments{
\item{x}{\code{gTrack} object}
}
\description{
Getting length of gTrack object gt
Usage:
length(gt)
}
\author{
Marcin Imielinski
}
