% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{lengths}
\alias{lengths}
\title{lengths}
\usage{
\S4method{lengths}{gTrack}(x)
}
\description{
gets lengths of data objects inside gTrack object gt
usage:
lengths(gt) # returns vector of lengths
}
\author{
Marcin Imielinski
}

