% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\name{lighten}
\alias{lighten}
\title{lighten}
\usage{
lighten(col, f)
}
\description{
lighten

lightens / darkens colors by brighness factor f (in -255 .. 255) that will make lighter if > 0 and darker < 0
}
\author{
Marcin Imielinski
}
\keyword{internal}

