% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.utils.R
\name{listify}
\alias{listify}
\title{convert input data into a list of length 'len' of type FUN}
\usage{
listify(x, FUN, len = 1)
}
\description{
convert input data into a list of length 'len' of type FUN
}
\keyword{internal}

