% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{mdata}
\alias{mdata}
\alias{mdata,gTrack-method}
\alias{mdata,gTrack,ANY,ANY,ANY-method}
\title{mdata}
\usage{
mdata(x, igr = NULL, jgr = NULL)

\S4method{mdata}{gTrack}(x, igr = NULL, jgr = igr)
}
\arguments{
\item{x}{\code{gTrack} object containing a matrix in the \code{mdata} slot}

\item{igr}{\code{GRegion} object specifying positions to subset matrix with}

\item{jgr}{\code{GRegion} object specifying positions to subset matrix with}
}
\description{
Accessing submatrices of mdata object for length 1 gTrack gt

Usage: (igr, jgr are GRanges objects corresponding to slices of matrix to be accessed from gt)
mdata(gt, igr, jgr)
}
\author{
Jeremiah Wala
}
