% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{mdata}
\alias{mdata}
\title{mdata}
\usage{
\S4method{mdata}{gTrack}(x, igr = NULL, jgr = igr)
}
\description{
Accessing submatrices of mdata object for length 1 gTrack gt

Usage: (igr, jgr are GRanges objects corresponding to slices of matrix to be accessed from gt)
mdata(gt, igr, jgr)
}
\author{
Jeremiah Wala
}

