% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\name{pencils}
\alias{pencils}
\title{pencils}
\usage{
pencils(x0, y0, x1, y1, leftpoint = 1, rightpoint = 1, lwd = 1,
  lty = par("lty"), col = "gray", border = col, col.head = col)
}
\description{
pencils

pencils = hexa-penta-rectagonal like pointy shapes often used to represent NGS read directionality (eg IGV)

parameterized like rect() function to represent horizontal (x0 and x1) and vertical (y0 and y1) extent of each object, however depending on the
value of rightpoint (0-1), leftpoint (0-1) will connect the top and bottom edges of the rectangle to the vertical midpoint of the rectangle
on the right / left side.  Values of 0 for leftpoint and rightpoint will yield a diamond, and values of 1 will yield a "classic" rectangle.

all arguments except leftpoint and rightpoint can be vectors (of length n) specifying n different barbs with different positions / parameters.
}
\author{
Marcin Imielinski
}
\keyword{internal}

