% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\name{plot.blank}
\alias{plot.blank}
\title{Make a blank plot}
\usage{
plot.blank(xlim = c(0, 1), ylim = c(0, 1), xlab = "", ylab = "",
  axes = F, ...)
}
\description{
Shortcut for making blank plot with no axes
}
\keyword{internal}

