% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{plot}
\alias{plot}
\title{plot}
\usage{
\S4method{plot}{gTrack,ANY}(x, windows = seqinfo2gr(seqinfo(x)),
  links = NULL, gap = NULL, y.heights = NULL, y.gaps = NULL,
  cex.xlabel = 1, cex.ylabel = 1, max.ranges = NA, links.feat = NULL,
  verbose = FALSE, ...)
}
\arguments{
\item{windows}{GRanges specifying windows to view (can also be GRangesList), default is whole genome}

\item{links}{GRangesList of signed locus pairs specifying links to draw above the plot,
optional GRangesList values meta data specify formatting of individual links:
$label (text label), $col (line color), $lwd (line weight), $lty (line style),
$arrow (arrow flag), $col.arrow (arrow color), $v (vertical spline bulge), $w (horizontal spline bulge)}

\item{gap}{scalar numeric specifying gap between windows (only relevant if windows has length>1). Units of the gap are in genome coordinates}

\item{max.ranges}{scalar numeric > 0 specifying max number of ranges to draw in a window (via sampling).  If specified, overrides gTrack max.ranges formatting feature.}

\item{...,}{additional last-minute formatting changes to the gtrack can be entered here (eg col = 'blue')}

\item{cex.tick}{vector specifying relative heights of gaps below and above (stacked) windows, only relevant if windows is GRangesList}
}
\description{
Plot gTrack object in multi-track genome browser view.  gTracks will appear as stacked interval
plots, line / bar / scatter plots, node-edge graphs, or triangle plots depending on the formatting
settings and additional data (eg mdata) provided.  gTracks can be drawn across several non-contiguous
"windows" of the genome

Additional argument "links" takes a GRangesList of signed interval pairs (ie each item is length 2)
representing genomic junctions as input. The junctions will be pointing right for + intervals and
left for - intervals.  The meta data of the links GRangesList ($lwd $col $lty) can be set
to modulate the color, witwidth, and line style of the junction links .

usage:
display(gt, win) # where win is a GRanges object
display(gt) # this will show the entire span of seqinfo(gt)
display(gt, '3:1e6-2e6') ## can use UCSC style strings
display(gt, GRanges(20, IRanges(20e6, 21e6)))
display(gt, '2', links = ra) # here, the entire chromosome 2 is shown, with rearrangements on top
}
\author{
Marcin Imielinski, Jeremiah Wala
}

