% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{reduce,gTrack-method}
\alias{reduce,gTrack-class}
\alias{reduce,gTrack-method}
\title{Computing the GRanges footprint of the gTrack object on the genome
usage:
reduce(gt) # outputs a GRanges}
\usage{
\S4method{reduce}{gTrack}(x, ...)
}
\arguments{
\item{...}{additional arguments to GRanges reduce function}
}
\description{
Computing the GRanges footprint of the gTrack object on the genome
usage:
reduce(gt) # outputs a GRanges
}

