% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{reduce}
\alias{reduce}
\alias{reduce,gTrack-method}
\title{reduce}
\usage{
\S4method{reduce}{gTrack}(x, ...)
}
\arguments{
\item{x}{\code{gTrack} object to retrieve reduced \code{GRanges} from}

\item{...}{additional arguments to GRanges reduce function}
}
\value{
\code{GRanges} with the minimal footprint of the \code{gTrack} data
}
\description{
Computing the GRanges footprint of the gTrack object on the genome
usage:
reduce(gt) # outputs a GRanges
}
\author{
Marcin Imielinski
}
