% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{seqinfo<-}
\alias{seqinfo,gTrack-method}
\alias{seqinfo<-}
\alias{seqinfo<-,gTrack-method}
\title{seqinfo,-}
\usage{
seqinfo(.Object) <- value

\S4method{seqinfo}{gTrack}(.Object) <- value
}
\description{
set seqinfo property of gTrack
}
\author{
Marcin Imielinski
}

