% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{seqinfo}
\alias{seqinfo}
\title{seqinfo}
\usage{
\S4method{seqinfo}{gTrack}(x)
}
\description{
returns Seqinfo of gTrack object gt
Usage:
seqinfo(gt)
}
\author{
Marcin Imielinski
}

