% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,gTrack-method}
\title{show}
\usage{
\S4method{show}{gTrack}(object)
}
\arguments{
\item{object}{\code{gTrack} to display}
}
\description{
Display a \code{gTrack} object
}
\author{
Marcin Imielinski
}
