% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,gTrack,ANY,ANY,ANY-method}
\title{[}
\usage{
\S4method{[}{gTrack,ANY,ANY,ANY}(x, i)
}
\description{
Subsetting tracks of gTrack gt
gt[1]
gt[ix] # where ix is an integer vector
}
\author{
Marcin Imielinski
}

