% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\name{track.gencode}
\alias{track.gencode}
\title{Constructor a \code{gTrack} from GENCODE transcripts}
\usage{
track.gencode(gencode = NULL, gene.collapse = TRUE, genes = NULL,
  grep = NULL, grepe = NULL, bg.col = alpha("blue", 0.1),
  cds.col = alpha("blue", 0.6), utr.col = alpha("purple", 0.4),
  st.col = "green", en.col = "red", grl.labelfield, gr.labelfield, col,
  cached = TRUE, cached.dir = Sys.getenv("GENCODE_DIR"),
  cached.path = paste(cached.dir, "gencode.composite.rds", sep = "/"),
  cached.path.collapsed = paste(cached.dir, "gencode.composite.collapsed.rds",
  sep = "/"), gr.srt.label = 0, gr.cex.label = 0.3, cex.label = 0.5,
  labels.suppress.gr = TRUE, drop.rp11 = TRUE, stack.gap = 1e+06, ...)
}
\arguments{
\item{gencode}{INFO INFO}

\item{gene.collapse}{scalar logical specifying whether to collapse genes by transcript (or used stored version of transcripts)}

\item{genes}{(optional) character vector specifying genes to limit gTrack object to}

\item{grep}{character vector for which to grep genes to positively select}

\item{grepe}{character vector for which to grep genes which to exclude}

\item{bg.col}{scalar character representing background color for genespan}

\item{cds.col}{scalar character representing background color for CDS}

\item{st.col}{scalar character representing color of CDS start}

\item{en.col}{scalar character representing color of CDS end}

\item{col}{INFO INFO INFO}

\item{cached}{logical scalar whether to use "cached" version provided with package}

\item{cached.dir}{INFO INFO INFO}

\item{cached.path}{INFO INFO INFO}

\item{cached.path.collapsed}{INFO INFO INFO}

\item{gr.srt.label}{scalar numeric specifying angle on exon label}

\item{gr.cex.label}{scalar numeric > 0 specifying character expansion on exon label}

\item{labels.suppress.gr}{boolean Flag specifying whether to suppress exon label plotting}

\item{drop.rp11}{boolean INFO INFO}

\item{stack.gap}{float stack.gap argument to gTrack}

\item{...}{additional arguments passed down to gTrack}

\item{grl.lablefield}{INFO INFO INFO}

\item{gr.lablefield}{INFO INFO INFO}
}
\description{
Returns gTrack object representing GENCODE transcripts and their components (utr, cds etc) with assigned colors.
Usually built from cached data objects but can also be built from provided GRangesList
}
\author{
Marcin Imielinski
}
