% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{$<-}
\alias{$<-}
\alias{$<-,gTrack-method}
\title{$<-}
\usage{
\S4method{$}{gTrack}(x, name) <- value
}
\arguments{
\item{x}{\code{gTrack} object to alter \code{formatting} field of}

\item{name}{\code{formatting} field to alter}

\item{value}{New value}
}
\description{
Setting formats of gTrack object - ie modifying the formatting(gt) data frame after
an object has already been instantiated

Usage:
gt$y.field = 'score'
gt$gr.colorfield[1] = 'readtype'
}
\author{
Marcin Imielinski
}
