% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\name{draw.triangle}
\alias{draw.triangle}
\title{draw.triangle}
\usage{
draw.triangle(grl, mdata = NA, y = NULL, ylim.parent = NULL,
  windows = NULL, win.gap = NULL, bg.col = "gray95", sigma = NA,
  col.min = "white", col.max = "red", palette.colors = NULL,
  cmap.min = NULL, cmap.max = NULL, msep.lwd = 1, legend.cex = 1,
  islog = FALSE)
}
\description{
internal draw.triangle function

Shortcut for making blank plot with no axes
}
\author{
Jeremiah Wala
}
\keyword{internal}

