% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\name{draw.triangle}
\alias{draw.triangle}
\title{draw.triangle}
\usage{
draw.triangle(grl, mdata, y, ylim.parent = NULL, windows = NULL,
  win.gap = NULL, m.bg.col, sigma = NA, col.min = "white",
  col.max = "red", gr.colormap = NA, cmap.min, cmap.max, m.sep.lwd,
  leg.params, min.gapwidth = 1, islog = FALSE)
}
\description{
internal draw.triangle function
}
\author{
Jeremiah Wala
}
\keyword{internal}

