% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{class}
\name{gTrack-class}
\alias{$,gTrack-class}
\alias{$,gTrack-method}
\alias{$<-,gTrack-class}
\alias{$<-,gTrack-method}
\alias{[,gTrack,ANY,ANY-method}
\alias{[,gTrack-class}
\alias{c,gTrack-class}
\alias{c,gTrack-method}
\alias{clear}
\alias{colormap}
\alias{colormap,gTrack-class}
\alias{colormap<-}
\alias{dat}
\alias{dat,gTrack-class}
\alias{dat,gTrack-method}
\alias{display}
\alias{display,gTrack-class}
\alias{edgs}
\alias{edgs,gTrack-class}
\alias{edgs,gTrack-method}
\alias{edgs<-}
\alias{edgs<-,gTrack-class}
\alias{formatting}
\alias{formatting,gTrack-class}
\alias{formatting,gTrack-method}
\alias{formatting<-}
\alias{formatting<-,gTrack-class}
\alias{formatting<-,gTrack-method}
\alias{gTrack}
\alias{gTrack-class}
\alias{lengths}
\alias{lengths,gTrack-class}
\alias{lengths,gTrack-method}
\alias{mdata,gTrack-method}
\alias{plot,gTrack,ANY-method}
\alias{plot,gTrack-class}
\alias{seqinfo,gTrack-class}
\alias{seqinfo,gTrack-method}
\alias{seqinfo<-}
\alias{seqinfo<-,gTrack-class}
\alias{seqinfo<-,gTrack-method}
\alias{show,gTrack-class}
\alias{show,gTrack-method}
\alias{vars}
\alias{vars,gTrack-class}
\alias{vars,gTrack-method}
\title{S4 class for \code{gTrack}}
\usage{
gTrack(...)

\S4method{[}{gTrack,ANY,ANY}(x, i)

\S4method{mdata}{gTrack}(x, igr = NULL, jgr = igr)

\S4method{$}{gTrack}(x, name)

\S4method{$}{gTrack}(x, name) <- value

\S4method{seqinfo}{gTrack}(x)

seqinfo(.Object) <- value

\S4method{seqinfo}{gTrack}(.Object) <- value

lengths(x)

\S4method{lengths}{gTrack}(x)

\S4method{c}{gTrack}(x, ..., recursive = FALSE)

formatting(.Object, ...)

\S4method{formatting}{gTrack}(.Object)

edgs(.Object)

\S4method{edgs}{gTrack}(.Object)

vars(.Object)

\S4method{vars}{gTrack}(.Object)

edgs(.Object) <- value

clear(.Object, ...)

dat(.Object, ...)

\S4method{dat}{gTrack}(.Object)

formatting(.Object) <- value

\S4method{formatting}{gTrack}(.Object) <- value

colormap(.Object, value)

colormap(.Object) <- value

\S4method{show}{gTrack}(object)

display(.Object, ...)

\S4method{plot}{gTrack,ANY}(x, y = seqinfo2gr(seinfo(x)), links = NULL,
  gap = NULL, y.heights = NULL, y.gaps = NULL, cex.xlabel = 1,
  cex.ylabel = 1, max.ranges = NA, links.feat = NULL, windows = NULL,
  verbose = FALSE, ...)
}
\arguments{
\item{...,}{additional last-minute formatting changes to the gtrack can be entered here (eg col = 'blue')}

\item{name}{vector or scalar character specifying name of this track, which will be displayed on label to the left of the track}

\item{links}{GRangesList of signed locus pairs specifying links to draw above the plot,
optional GRangesList values meta data specify formatting of individual links:
$label (text label), $col (line color), $lwd (line weight), $lty (line style),
$arrow (arrow flag), $col.arrow (arrow color), $v (vertical spline bulge), $w (horizontal spline bulge)}

\item{gap}{scalar numeric specifying gap between windows (only relevant if windows has length>1). Units of the gap are in genome coordinates}

\item{max.ranges}{vector or scalar numeric specifying what is the max number of ranges to plot in a window (formatting)}

\item{windows}{GRanges specifying windows to view (can also be GRangesList), default is whole genome}

\item{data}{This is an instance of one of the following objects: (1) GRanges (2) GRangesList (3) ffTrack or
(4) character representing Bed, Wig, BigWig, or .rds GRanges file.  It can also be a list of the above (if specifying multiplying
 tracks.}

\item{y.field}{vector or scalar character specifying meta data field of GRanges or GRangesList to use for "y axis" coordinate when plotting
numeric tracks, (note: for a RleList or ffTrack this field is automatically set to "score"), default is NA for non-numeric tracks}

\item{height}{vector or scalar numeric specifying height of track(s) (in relative units)}

\item{gr.labelfield}{vector or scalar character specifying which GRanges meta data field to use for GRanges label (default "label") (formatting)}

\item{grl.labelfield}{vector or scalar character specifying which GRanges meta data field to use for GRangesList label (default "label") (formatting)}

\item{legend}{vector or scalar logical specifying whether to draw a legend for this track (formatting)}

\item{legend.xpos}{vector or scalar numeric between 0 and 1 specifying what relative x position in the plot to place the legend for this track  (formatting)}

\item{legend.ypos}{vector or scalar numeric between 0 and 1 specifying what relative y position in the plot to place the legend for this track  (formatting)}

\item{legend.ncol}{vector or scalar positive integer specifying how many columns to put in legend (formatting)}

\item{legend.xjust}{vector or scalar of {0, 1, 2} specifying x justification of legend (formatting)}

\item{legend.yjust}{vector or scalar of {0, 1, 2} specifying y justification of legend (formatting)}

\item{legend.maxitems}{vector or scalar positive integer specifying what is the maximum number of items to include in legend  (formatting)}

\item{label.suppress}{vector or scalar logical flag specifying whether to suppress all GRanges / GRangesList label drawing  (formatting)}

\item{label.suppress.gr}{vector or scalar logical flag specifying whether to suppress GRanges label drawing  (formatting)}

\item{label.suppress.grl}{vector or scalar logical flag specifying whether to suppress GRangesList label drawing  (formatting)}

\item{ygap}{vector or scalar numeric specifying gap between tracks}

\item{stack.gap}{vector or scalar.numeric specifying x gap between stacking non-numeric GRanges or GrangesLists items in track(s)}

\item{ywid}{vector or scalar numeric specifying the y-extent of individual ranges (in local plot coordinates)}

\item{col}{vector or scalar character specifying static color for track(s), if NA then color is specified by colormaps() property or gr.colorfield or col meta data field of GRanges / GRangesList data object}

\item{border}{vector or scalar character specifying static border for polygons in track(s), if NA then $border is determine duing gr.colorfield / colormap or meta field $border of GRanges / GRangesList}

\item{angle}{vector of scalar numeric specifying angle of polygons that represent signed range'}

\item{lift}{vector or scalar logical flag specifying whether to lift this track to other chainedTrack
items (only relevant if used wihtin chainedTracks object)}

\item{split}{vector or scalar logical flag specifying whether to split when lifting (only relevant if used wihtin chainedTracks object)}

\item{colormaps}{length(.Object) length named list of named vectors whose entry i maps uniques value of a data field to colors.  The data.field is specified by the name of the list entry, and the unique values / colors are specified by the named vector.}

\item{edges}{Data frame of columns $from, $to, and optional fields $col, $lwd, and $lty, specifying edges linking data items.
Also can be a list of the above if specifying multiple tracks (and must be compatible in length with data arg)}

\item{xaxis.suffix}{vector or scalar numeric specifying the suffix that will be used in describin x axis coordinates (TODO: move to display) (formatting)}

\item{xaxis.unit}{vector or scalar numeric specifying the unit that will be used in describing x axis coordinates (TODO: move to display)  (formatting)}

\item{xaxis.round}{vector or scalar non-neg integer specifying number of decimals to round xaxis coordinate labels (formatting)}

\item{xaxis.nticks}{vector or scalar positive integer specifying how many xaxis ticks to optimally draw (formatting)}

\item{xaxis.label.angle}{vector or scalar numeric between 0 and 360 specifying angle with which to draw xaxis coordinate labels (formatting)}

\item{xaxis.newline}{vector or scalar logical specifying whether to draw a newline in the xaxis coordinate labels (formatting)}

\item{lwd.border}{vector or scalar integer specifying the thickness of the polygon borders (formatting)}

\item{cex.label}{vector or scalar numeric specifying the expansion factor of the range labels (formatting)}

\item{hadj.label}{vector or scalar numeric specifying the horizontal adjustment of the range labels (formatting)}

\item{vadj.label}{vector or scalar numeric specifying the vertical adjustment of the range labels (formatting)}

\item{ypad}{vector or scalar numeric between 0 and 1 specifying how much whitespace padding to add within panel (formatting)}

\item{circles}{vector or scalar logical specifying whether to scatter plot range data (formatting)}

\item{lines}{vector or scalar logical specifying whether to line plot range data (formatting)}

\item{bars}{vector or scalar logical specifying whether to bar plot range data (formatting)}

\item{y0.bar}{vector or scalar numeric specifying where to draw the lower boundary of a bar in a bar plot (only applicable if bars == T) (formatting)}

\item{source.file.chrsub}{vector or scalar logical specifying whether or not sub "chr" out of any external files (e.g. UCSC style files) (formatting)}

\item{y.grid.col}{vector or scalar character specifying color of "gridlines" used to specify numeric track data (formatting)}

\item{y.grid.cex}{vector or scalar non-neg numeric specifying character expansion for y tick / y grid labels (formatting)}

\item{y.grid.lty}{vector or scalar positive integer specifying line style of y grid lines for numeric tracks (formatting)}

\item{y.grid.lwd}{vector or scalar positive integer specifying thickness of y grid lines for numeric tracks (formatting)}

\item{y.grid.labx}{vector or scalar positive integer specifying fraction of xlim left of plot to place y axis labels (formatting)}

\item{yaxis}{vector or scalar logical specifying whether to print yaxis (formatting)}

\item{yaxis.pretty}{vector or scalar positive integer specifying how many ticks to optimally draw on yaxis (formatting)}

\item{draw.var}{vector or scalar logical specifying whether to draw.var for GRanges / GRangesList  specifying reads
(GRanges must contain $cigar +/- $MD field) (formatting)}

\item{draw.paths}{vector or scalar logical specifying whether to interpret GRangesLists as "paths" and connect them with a
a set of spline curves.  (formatting)}

\item{path.col}{vector or scalar character specifying color of path (only applicable for tracks in which draw.paths = T) (formatting)}

\item{path.col.arrow}{vector or scalar character specifying color of arrow of path (only applicable for tracks in which draw.paths = T) (formatting)}

\item{path.cex.arrow}{vector or scalar numeric > 0 specifying expansion factor of arrow of path (only applicable for tracks
in which draw.paths = T)  (formatting)}

\item{path.stack.y.gap}{vector or scalar numeric > 0 specifying y stack gap of paths (only applicable for tracks in which draw.paths = T)  (formatting)}

\item{path.stack.x.gap}{vector or scalar numeric > 0 specifying x stack gap for paths  (only applicable for tracks in which draw.paths = T)  (formatting)}

\item{path.cex.v}{vector or scalar numeric > 0 specifying vertical bulge of sline in paths (only applicable for tracks in which draw.paths = T)  (formatting)}

\item{path.cex.h}{vector or scalar numeric > 0 specifying horizontal bulge of spline in paths (formatting)
(only applicable for tracks in which draw.paths = T)  (formatting)}

\item{draw.backbone}{vector or scalar logical specifying whether to draw "backbone" connecting different items in a GRangesList item (formatting)}

\item{cex.tick}{vector or scalar numeric specifying expansion factor for axis tick labels  (formatting)}

\item{cex.tick.len}{vector or scalar numeric specifying lengths for axis ticks  (formatting)}

\item{gr.cex.label}{vector or scalar numeric specifying GRanges label character expansion (default is cex.label)  (formatting)}

\item{gr.srt.label}{vector or scalar numeric between 0 and 180 specifying rotation of GRanges labels  (formatting)}

\item{sep.lty}{vector or scalar integer specifying line style for window separators  (formatting)}

\item{sep.lwd}{vector or scalar numeric specifying line thickness for window separators  (formatting)}

\item{cmap.min}{minimum saturating data value of color map for triangle plot}

\item{cmap.max}{maximum saturating data value of color map for triangle plot}

\item{cex.tick}{vector specifying relative heights of gaps below and above (stacked) windows, only relevant if windows is GRangesList}

\item{max.ranges}{scalar numeric > 0 specifying max number of ranges to draw in a window (via sampling).  If specified, overrides gTrack max.ranges formatting feature.}
}
\description{
Class \code{gTrack} defines a subsettable object that wraps formatting information around
genomic data (GRanges, GRangesList, RleLists, UCSC formats, ffTrack) and can be displayed
in a "genome browser" style plot.

Arguments described as "formatting" are vectors.  They are replicated (if necessary) to match the length of the object.

Subsetting tracks of gTrack gt
gt[1]
gt[ix] # where ix is an integer vector

Usage: (igr, jgr are GRanges objects corresponding to slices of matrix to be accessed from gt)
mdata(gt, igr, jgr)

Accessing columns of gTrack formatting data.frame

Usage:
gt$y.field = 'score'
gt$gr.colorfield[1] = 'readtype'

returns Seqinfo of gTrack object gt
Usage:
seqinfo(gt)



gets lengths of data objects inside gTrack object gt
usage:
lengths(gt) # returns vector of lengths

Concatenate gTrack objects gt1, gt2, gt3
usage:
c(gt1, gt2, gt3) # returns a gTrack object with the component tracks "stacked"

Get data frame specifying formatting of gTrack object gt
usage:
formatting(gt)
gt # will just display the formatting data.frame

usage:
edgs(gt)



usage
dat(gt)

Formatting

usage:
#gt is a gTrack object
formatting(td)$height <- 2

Additional argument "links" takes a GRangesList of signed interval pairs (ie each item is length 2)
representing genomic junctions as input. The junctions will be pointing right for + intervals and
left for - intervals.  The meta data of the links GRangesList ($lwd $col $lty) can be set
to modulate the color, witwidth, and line style of the junction links .
}
\details{
#can also just use $ directly
td$height <- 2

usage:
display(gt, win) # where win is a GRanges object
display(gt) # this will show the entire span of seqinfo(gt)
display(gt, '3:1e6-2e6') ## can use UCSC style strings
display(gt, GRanges(20, IRanges(20e6, 21e6)))
display(gt, '2', links = ra) # here, the entire chromosome 2 is shown, with rearrangements on top
}
\section{Slots}{

\describe{
  \item{data}{length(.Object) length list containing genomic data (e.g. GRanges, GrangesLists, RleLists, path to ucsc file or ffTrack file on disk)}
  \item{seqinfo}{Seqinfo object}
  \item{colormap}{length(.Object) length named list of named vectors whose entry i specifies the colormap for the meta data field of object entry i (specified by the name) and maps unique value of that data field to colors (specified by the named vector)}
  \item{edges}{list of data.frames of length length(.Object) which has columns $from, $to, and optional fields $col, $lwd, and $lty to specify splined edges joining data items in the corresponding track}
}
}

