% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{class}
\name{gTrack-class}
\alias{gTrack}
\alias{gTrack-class}
\title{Construct a new \code{gTrack}}
\usage{
gTrack(data = NULL, xaxis = gaxis(), yaxis = gaxis(),
  legend = glegend(), path = gpath(), separator = gsep(),
  label = glabel(), y.field = NA, mdata = NULL, edges = NULL,
  vars = NULL, colormaps = NULL, height = 10, ygap = 2, stack.gap = 0,
  col = NA, border = "black", angle = 15, name = "",
  gr.colorfield = NA, lwd.border = 1, smooth = NA, round = NA,
  ywid = NA, ypad = 0, seqinfo = NA, plot.type = "ranges",
  draw.paths = FALSE, draw.var = FALSE, triangle = !is.null(mdata),
  gr.labelfield = NA, max.ranges = 50000, chr.sub = TRUE, edgevars = NA,
  cmap.min = NA, cmap.max = NA, bg.col = "white", formatting = NA)
}
\arguments{
\item{data}{Instance of one of the following objects: (1) GRanges (2) GRangesList (3) ffTrack or
(4) character representing Bed, Wig, BigWig, or .rds GRanges file.  It can also be a list of the above (if specifying multiplying
 tracks.}

\item{y.field}{vector or scalar character referencing meta data field of GRanges or GRangesList to use for "y axis" coordinate when plotting
numeric tracks, (note: for a RleList or ffTrack this field is automatically set to "score"), default is NA for non-numeric tracks}

\item{mdata}{Square matrix representing values at connections between genomic elements (e.g. chr8:1,000-2,000 to chr3:5,000-6,000). The number of
rows (and thus columns) must be equal to the length of the \code{data} element. The value of the (i,j) element of this matrix represents a value
for the intersection of \code{data[i]} and \code{data[j]}. Note that all connections are assumed to be symmetric, and only the upper triangle
will be read. Displays a \code{warning} if the matrix is non-symmetric}

\item{edges}{Data frame of columns $from, $to, and optional fields $col, $lwd, and $lty, specifying edges linking data items.
Also can be a list of the above if specifying multiple tracks (and must be compatible in length with data arg)}

\item{colormaps}{length(.Object) length named list of named vectors whose entry i maps uniques value of a data field to colors.  The data.field is specified by the name of the list entry, and the unique values / colors are specified by the named vector.}

\item{height}{height of track(s) (in relative units)}

\item{ygap}{vector or scalar numeric specifying gap between tracks}

\item{stack.gap}{vector or scalar.numeric specifying x gap between stacking non-numeric GRanges or GrangesLists items in track(s)}

\item{col}{vector or scalar character specifying static color for track(s), if NA then color is specified by colormaps() property or gr.colorfield or col meta data field of GRanges / GRangesList data object}

\item{border}{vector or scalar character specifying static border for polygons in track(s), if NA then $border is determine duing gr.colorfield / colormap or meta field $border of GRanges / GRangesList}

\item{angle}{vector of scalar numeric specifying angle of polygons that represent signed range'
items (only relevant if used within chainedTracks object)}

\item{name}{name of this track, which will be displayed on label to the left of the track}

\item{lwd.border}{vector or scalar integer specifying the thickness of the polygon borders (formatting)}

\item{ywid}{vector or scalar numeric specifying the y-extent of individual ranges (in local plot coordinates)}

\item{ypad}{vector or scalar numeric between 0 and 1 specifying how much whitespace padding to add within panel (formatting)}

\item{plot.type}{range plot (\code{ranges}), scatter plot (\code{scatter}), line plot (\code{line}), or bar plot (\code{bars})}

\item{draw.paths}{vector or scalar logical specifying whether to interpret GRangesLists as "paths" and connect them with a
a set of spline curves.  (formatting)}

\item{draw.var}{vector or scalar logical specifying whether to draw.var for GRanges / GRangesList  specifying reads
(GRanges must contain $cigar +/- $MD field) (formatting)}

\item{max.ranges}{vector or scalar numeric specifying what is the max number of ranges to plot in a window (formatting)}

\item{cmap.min}{minimum saturating data value of color map for triangle plot}

\item{cmap.max}{maximum saturating data value of color map for triangle plot}

\item{split}{vector or scalar logical flag specifying whether to split when lifting (only relevant if used wihtin chainedTracks object)}

\item{draw.backbone}{vector or scalar logical specifying whether to draw "backbone" connecting different items in a GRangesList item (formatting)}

\item{...}{additional formatting arguments to gTrack}
}
\description{
S4 class for \code{gTrack}

Class \code{gTrack} defines a subsettable object that wraps formatting information around
genomic data (GRanges, GRangesList, RleLists, UCSC formats, ffTrack) and can be displayed
in a "genome browser" style plot.

Arguments described as "formatting" are vectors.  They are replicated (if necessary) to match the length of the object.
}
\section{Slots}{

\describe{
  \item{data}{length(.Object) length list containing genomic data (e.g. GRanges, GrangesLists, RleLists, path to ucsc file or ffTrack file on disk)}
  \item{seqinfo}{Seqinfo object}
  \item{colormap}{length(.Object) length named list of named vectors whose entry i specifies the colormap for the meta data field of object entry i (specified by the name) and maps unique value of that data field to colors (specified by the named vector)}
  \item{edges}{list of data.frames of length length(.Object) which has columns $from, $to, and optional fields $col, $lwd, and $lty to specify splined edges joining data items in the corresponding track}
  \item{formatting}{\code{data.frame} holding all of the formatting options}
  \item{mdata}{\code{matrix} holding interaction data between genomic loci (e.g. hiC)}
  \item{vars}{List of variants associated with track}
  \item{xaxis}{gaxis object holding all of the xaxis formatting data}
  \item{yaxis}{list of gaxis object holding all of the yaxis formatting data}
}
}
\author{
Marcin Imielinski

Marcin Imielinski
}

