% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{class}
\name{gTrack-class}
\alias{gTrack-class}
\alias{gTrack}
\title{Construct a new \code{gTrack}}
\usage{
gTrack(data = NULL, y.field = NA, mdata = NULL, edges = NULL,
  vars = NULL, colormaps = NULL, height = 10, ygap = 2, ylab = "",
  stack.gap = 0, cex.label = 1, gr.cex.label = cex.label * 0.8,
  gr.srt.label = 0, col = NA, border = NA, angle = 15, name = "",
  gr.colorfield = NA, y.quantile = 0.01, y.cap = T, lwd.border = 1,
  hadj.label = 1, vadj.label = 0.5, smooth = NA, round = NA,
  ywid = NA, ypad = 0, seqinfo = NA, circles = FALSE, lines = FALSE,
  bars = FALSE, draw.paths = FALSE, draw.var = FALSE,
  triangle = !is.null(mdata), max.ranges = 50000, source.file.chrsub = T,
  y0.bar = NA, yaxis = !is.na(y.field), yaxis.pretty = 5, yaxis.cex = 1,
  chr.sub = TRUE, edgevars = NA, gr.labelfield = NA,
  grl.labelfield = NA, xaxis.prefix = "", xaxis.unit = 1,
  xaxis.suffix = "", xaxis.round = 3, xaxis.cex.label = 1,
  xaxis.newline = TRUE, xaxis.chronly = FALSE, xaxis.width = TRUE,
  xaxis.interval = "auto", xaxis.label.angle = 0, xaxis.ticklen = 1,
  xaxis.cex.tick = 1, sep.lty = 2, sep.lwd = 1, sep.bg.col = "gray95",
  sep.draw = TRUE, y0 = NA, y1 = NA, m.sep.lwd = 1, m.bg.col = NA,
  cmap.min = NA, cmap.max = NA, labels.suppress = FALSE,
  labels.suppress.grl = labels.suppress,
  labels.suppress.gr = labels.suppress, bg.col = "white", formatting = NA)
}
\arguments{
\item{data}{Instance of one of the following objects: (1) GRanges (2) GRangesList (3) ffTrack or
(4) character representing Bed, Wig, BigWig, or .rds GRanges file.  It can also be a list of the above (if specifying multiplying
 tracks.}

\item{y.field}{vector or scalar character referencing meta data field of GRanges or GRangesList to use for "y axis" coordinate when plotting
numeric tracks, (note: for a RleList or ffTrack this field is automatically set to "score"), default is NA for non-numeric tracks}

\item{mdata}{Square matrix representing values at connections between genomic elements (e.g. chr8:1,000-2,000 to chr3:5,000-6,000). The number of
rows (and thus columns) must be equal to the length of the \code{data} element. The value of the (i,j) element of this matrix represents a value
for the intersection of \code{data[i]} and \code{data[j]}. Note that all connections are assumed to be symmetric, and only the upper triangle
will be read. Displays a \code{warning} if the matrix is non-symmetric}

\item{edges}{Data frame of columns $from, $to, and optional fields $col, $lwd, and $lty, specifying edges linking data items.
Also can be a list of the above if specifying multiple tracks (and must be compatible in length with data arg)}

\item{colormaps}{length(.Object) length named list of named vectors whose entry i maps uniques value of a data field to colors.  The data.field is specified by the name of the list entry, and the unique values / colors are specified by the named vector.}

\item{height}{vector or scalar numeric specifying height of track(s) (in relative units)}

\item{ygap}{vector or scalar numeric specifying gap between tracks}

\item{stack.gap}{vector or scalar.numeric specifying x gap between stacking non-numeric GRanges or GrangesLists items in track(s)}

\item{cex.label}{size of top level label (e.g. GRAngesList item)}

\item{gr.cex.label}{vector or scalar numeric specifying GRanges label character expansion (default is cex.label)  (formatting)}

\item{gr.srt.label}{rotation of GRanges label}

\item{col}{vector or scalar character specifying static color for track(s), if NA then color is specified by colormaps() property or gr.colorfield or col meta data field of GRanges / GRangesList data object}

\item{border}{vector or scalar character specifying static border for polygons in track(s), if NA then $border is determine duing gr.colorfield / colormap or meta field $border of GRanges / GRangesList}

\item{angle}{vector of scalar numeric specifying angle of polygons that represent signed range'
items (only relevant if used within chainedTracks object)}

\item{name}{vector or scalar character specifying name of this track, which will be displayed on label to the left of the track}

\item{lwd.border}{vector or scalar integer specifying the thickness of the polygon borders (formatting)}

\item{hadj.label}{vector or scalar numeric specifying the horizontal adjustment of the range labels (formatting)}

\item{vadj.label}{vector or scalar numeric specifying the vertical adjustment of the range labels (formatting)}

\item{ywid}{vector or scalar numeric specifying the y-extent of individual ranges (in local plot coordinates)}

\item{ypad}{vector or scalar numeric between 0 and 1 specifying how much whitespace padding to add within panel (formatting)}

\item{circles}{vector or scalar logical specifying whether to scatter plot range data (formatting)}

\item{lines}{vector or scalar logical specifying whether to line plot range data (formatting)}

\item{bars}{vector or scalar logical specifying whether to bar plot range data (formatting)}

\item{draw.paths}{vector or scalar logical specifying whether to interpret GRangesLists as "paths" and connect them with a
a set of spline curves.  (formatting)}

\item{draw.var}{vector or scalar logical specifying whether to draw.var for GRanges / GRangesList  specifying reads
(GRanges must contain $cigar +/- $MD field) (formatting)}

\item{max.ranges}{vector or scalar numeric specifying what is the max number of ranges to plot in a window (formatting)}

\item{source.file.chrsub}{vector or scalar logical specifying whether or not sub "chr" out of any external files (e.g. UCSC style files) (formatting)}

\item{y0.bar}{vector or scalar numeric specifying where to draw the lower boundary of a bar in a bar plot (only applicable if bars == T) (formatting)}

\item{yaxis}{vector or scalar logical specifying whether to print yaxis (formatting)}

\item{yaxis.pretty}{vector or scalar positive integer specifying how many ticks to optimally draw on yaxis (formatting)}

\item{gr.labelfield}{vector or scalar character specifying which GRanges meta data field to use for GRanges label (default "label") (formatting)}

\item{grl.labelfield}{vector or scalar character specifying which GRanges meta data field to use for GRangesList label (default "label") (formatting)}

\item{xaxis.unit}{vector or scalar numeric specifying the unit that will be used in describing x axis coordinates (TODO: move to display)  (formatting)}

\item{xaxis.suffix}{vector or scalar numeric specifying the suffix that will be used in describin x axis coordinates (TODO: move to display) (formatting)}

\item{xaxis.round}{vector or scalar non-neg integer specifying number of decimals to round xaxis coordinate labels (formatting)}

\item{xaxis.newline}{vector or scalar logical specifying whether to draw a newline in the xaxis coordinate labels (formatting)}

\item{xaxis.width}{Logical scalar specifying whether to add window width to xaxis window labels [TRUE]}

\item{xaxis.label.angle}{vector or scalar numeric between 0 and 360 specifying angle with which to draw xaxis coordinate labels (formatting)}

\item{xaxis.ticklen}{Scalar numeric specifying lengths for axis ticks  (formatting)}

\item{xaxis.cex.tick}{Scalar numeric specifying expansion factor for axis tick labels  (formatting)}

\item{sep.lty}{Scalar integer specifying line style for window separators [2] (dashed line)}

\item{sep.lwd}{Scalar numeric specifying line thickness for window separators [1]}

\item{sep.bg.col}{Color (supplied by name or hex code) for the background of the windows [gray95]}

\item{sep.draw}{Logical allowing separators to be turned off [FALSE]}

\item{cmap.min}{minimum saturating data value of color map for triangle plot}

\item{cmap.max}{maximum saturating data value of color map for triangle plot}

\item{labels.suppress}{whether to suppress labels (both GRangesList and GRanges)}

\item{labels.suppress.gr}{whether to suppress GRanges labels}

\item{legend.maxitems}{Scalar positive integer specifying what is the maximum number of items to include in legend [Inf]}

\item{label.suppress}{vector or scalar logical flag specifying whether to suppress all GRanges / GRangesList label drawing  (formatting)}

\item{label.suppress.gr}{vector or scalar logical flag specifying whether to suppress GRanges label drawing  (formatting)}

\item{label.suppress.grl}{vector or scalar logical flag specifying whether to suppress GRangesList label drawing  (formatting)}

\item{cex.label.gr}{size of bottom level label (i.e. GRanges in a GrangesListitem, e.g. exons in a gene)}

\item{split}{vector or scalar logical flag specifying whether to split when lifting (only relevant if used wihtin chainedTracks object)}

\item{xaxis.nticks}{vector or scalar positive integer specifying how many xaxis ticks to optimally draw (formatting)}

\item{cex.label}{vector or scalar numeric specifying the expansion factor of the range labels (formatting)}

\item{y.grid.col}{vector or scalar character specifying color of "gridlines" used to specify numeric track data (formatting)}

\item{y.grid.cex}{vector or scalar non-neg numeric specifying character expansion for y tick / y grid labels (formatting)}

\item{y.grid.lty}{vector or scalar positive integer specifying line style of y grid lines for numeric tracks (formatting)}

\item{y.grid.lwd}{vector or scalar positive integer specifying thickness of y grid lines for numeric tracks (formatting)}

\item{y.grid.labx}{vector or scalar positive integer specifying fraction of xlim left of plot to place y axis labels (formatting)}

\item{path.col}{vector or scalar character specifying color of path (only applicable for tracks in which draw.paths = T) (formatting)}

\item{path.col.arrow}{vector or scalar character specifying color of arrow of path (only applicable for tracks in which draw.paths = T) (formatting)}

\item{path.cex.arrow}{vector or scalar numeric > 0 specifying expansion factor of arrow of path (only applicable for tracks
in which draw.paths = T)  (formatting)}

\item{path.stack.y.gap}{vector or scalar numeric > 0 specifying y stack gap of paths (only applicable for tracks in which draw.paths = T)  (formatting)}

\item{path.stack.x.gap}{vector or scalar numeric > 0 specifying x stack gap for paths  (only applicable for tracks in which draw.paths = T)  (formatting)}

\item{path.cex.v}{vector or scalar numeric > 0 specifying vertical bulge of sline in paths (only applicable for tracks in which draw.paths = T)  (formatting)}

\item{path.cex.h}{vector or scalar numeric > 0 specifying horizontal bulge of spline in paths (formatting)
(only applicable for tracks in which draw.paths = T)  (formatting)}

\item{draw.backbone}{vector or scalar logical specifying whether to draw "backbone" connecting different items in a GRangesList item (formatting)}

\item{gr.srt.label}{vector or scalar numeric between 0 and 180 specifying rotation of GRanges labels  (formatting)}

\item{labels.suppress.gr}{whether to suppress GRangesList labels}

\item{...}{additional formatting arguments to gTrack}
}
\description{
S4 class for \code{gTrack}

Class \code{gTrack} defines a subsettable object that wraps formatting information around
genomic data (GRanges, GRangesList, RleLists, UCSC formats, ffTrack) and can be displayed
in a "genome browser" style plot.

Arguments described as "formatting" are vectors.  They are replicated (if necessary) to match the length of the object.
}
\section{Slots}{

\describe{
  \item{data}{length(.Object) length list containing genomic data (e.g. GRanges, GrangesLists, RleLists, path to ucsc file or ffTrack file on disk)}
  \item{seqinfo}{Seqinfo object}
  \item{colormap}{length(.Object) length named list of named vectors whose entry i specifies the colormap for the meta data field of object entry i (specified by the name) and maps unique value of that data field to colors (specified by the named vector)}
  \item{edges}{list of data.frames of length length(.Object) which has columns $from, $to, and optional fields $col, $lwd, and $lty to specify splined edges joining data items in the corresponding track}
  \item{formatting}{\code{data.frame} holding all of the formatting options}
  \item{mdata}{\code{matrix} holding interaction data between genomic loci (e.g. hiC)}
  \item{vars}{List of variants associated with track}
}
}

\author{
Marcin Imielinski

Marcin Imielinski
}
