% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\name{karyogram}
\alias{karyogram}
\title{karyogram}
\usage{
karyogram(file = NULL, hg19 = TRUE, bands = TRUE, arms = TRUE,
  tel.width = 2e+06, ...)
}
\arguments{
\item{hg19}{logical scalar flag, if T returns gTrack for hg19}

\item{bands}{logical scalar, if T returns gTrack with colored giemsa bands}

\item{arms}{logical scalar, if T and bands F, returns chromosome arms with different colors and centromeres and telomeres marked, if arms is F and bands F returns chromosomes, each with a different color}

\item{tel.width}{numeric scalar, specifies telomere width in bases (only relevant if arms = T, bands = F)}

\item{...}{Additional arguments sent to the \code{gTrack} constructor}
}
\description{
Returns gTrack displaying band pattern for hg18 or hg19 karyotype
}
\author{
Marcin Imielinski
}
