% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{lengths}
\alias{lengths}
\alias{lengths,gTrack-method}
\title{lengths}
\usage{
\S4method{lengths}{gTrack}(x)
}
\arguments{
\item{x}{\code{gTrack} to get the lengths from}
}
\description{
gets lengths of data objects inside gTrack object gt
usage:
lengths(gt) # returns vector of lengths
}

