% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\name{plot.blank}
\alias{plot.blank}
\title{plot.blank}
\usage{
\method{plot}{blank}(xlim = c(0, 1), ylim = c(0, 1), xlab = "",
  ylab = "", axes = F, ...)
}
\description{
Make a blank plot

Shortcut for making blank plot with no axes
}
\author{
Marcin Imielinski
}
\keyword{internal}

