% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\name{readData}
\alias{readData}
\title{utility function to load data files into variables}
\usage{
readData(..., environment = NULL)
}
\description{
utility function to load data files into variables
}
\keyword{internal}

