% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{sep}
\alias{sep}
\alias{sep,gTrack-method}
\title{Retrieves the seperator graphical parameters}
\format{field responsible for
formatting the windows and their separators}
\usage{
sep(.Object)

\S4method{sep}{gTrack}(.Object)
}
\arguments{
\item{.Object}{\code{gTrack} object to extract separator params from}
}
\value{
\code{data.frame} with the seperator paramters
}
\description{
Return the portion of the gTrack
}
\author{
Jeremiah Wala
}
