% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{[}
\alias{[}
\title{[}
\usage{
\S4method{[}{gTrack,ANY,ANY,ANY}(x, i)
}
\description{
Subsetting tracks of gTrack gt
gt[1]
gt[ix] # where ix is an integer vector
}
\author{
Marcin Imielinski
}

