% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gTrack.R
\name{track.splice}
\alias{track.splice}
\alias{track.splice,gTrack-class}
\title{Given set of exons and rna bam (eg from tophat) determines junction and exon read density and returns a gTrack object
of splicing graph}
\usage{
track.splice(ex = NULL, region = NULL, bam, verbose = TRUE,
  infer.exons = FALSE, min.reads = 0, min.exon.width = 10,
  max.exon.width = 1000)
}
\arguments{
\item{ex}{GRanges of candidate exons}

\item{bam}{path to indexed RNA seq bam}

\item{verbose}{}
}
\description{
Given set of exons and rna bam (eg from tophat) determines junction and exon read density and returns a gTrack object
of splicing graph
}

