% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gTrack.R
\docType{methods}
\name{xaxis}
\alias{xaxis}
\alias{xaxis,gTrack-method}
\title{Retrieves the xaxis parameters}
\usage{
xaxis(.Object)

\S4method{xaxis}{gTrack}(.Object)
}
\arguments{
\item{.Object}{\code{gTrack} object to retrieve xaxis parameters from}
}
\value{
\code{data.frame} which is a subset of \code{formatting} showing xaxis params
}
\description{
Return the portion of the gTrack @format field responsible for
formatting the x-axis
}
\author{
Jeremiah Wala
}
